pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "binance-connector>=3.12.0",
    "ccxt>=4.4.63",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pybit>=5.9.0",
    "python-okx>=0.3.5",
    "requests>=2.32.3",
    "streamlit>=1.42.2",
    "twilio>=9.4.6",
    "urllib3>=2.3.0",
]
